Creating Lexer and Tokens
I have created a reader that goes through every character of my code and decides what it is. Is it an Integer or a Float? Is it Multiplication or Addition etc. I have also worked on errors and error handling. If an incorrect character is inputted, it will return with a printed error line.